@@ -1,4 +1,6 @@
+# frozen_string_literal: true
+
 class User
-  include ChargebeeRails::Subscriber
+  include ChargebeeRails::Customer
   attr_accessor :chargebee_id, :first_name, :last_name, :email, :phone
 end
