@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require File.dirname(__FILE__) + '/../lib/chargebee_rails'
 require 'rspec'
 require 'webmock/rspec'
@@ -7,41 +9,19 @@ require 'pry'
 require 'chargebee'
 
 RSpec.configure do |config|
-  # rspec-expectations config goes here. You can use an alternate
-  # assertion/expectation library such as wrong or the stdlib/minitest
-  # assertions if you prefer.
-  config.expect_with :rspec do |expectations|
-    # This option will default to `true` in RSpec 4. It makes the `description`
-    # and `failure_message` of custom matchers include text for helper methods
-    # defined using `chain`, e.g.:
-    # be_bigger_than(2).and_smaller_than(4).description
-    #   # => "be bigger than 2 and smaller than 4"
-    # ...rather than:
-    #   # => "be bigger than 2"
-    expectations.include_chain_clauses_in_custom_matcher_descriptions = true
-  end
-
-  # default, enables both `should` and `expect`
-  config.expect_with(:rspec) { |c| c.syntax = [:should, :expect] }
-
   # rspec-mocks config goes here. You can use an alternate test double
   # library (such as bogus or mocha) by changing the `mock_with` option here.
   config.mock_with :mocha
-  
 end
 
 VCR.configure do |config|
-  config.cassette_library_dir = "spec/support/vcr_cassettes"
+  config.cassette_library_dir = 'spec/support/vcr_cassettes'
   config.hook_into :webmock
   config.configure_rspec_metadata!
-  config.default_cassette_options = { record: :new_episodes }
   config.around_http_request do |request|
     VCR.use_cassette('chargebee_apis', record: :new_episodes, &request)
   end
 end
 
-
-ChargeBee.configure(
-  api_key: "dummy-api-key", 
-  site: "dummy-site"
-)
+ChargeBee.configure(site: 'dummy-site',
+                    api_key: "dummy-api-key")
